Focus the schedule dialog box when it pops up
Gives the schedule dialog box a tabIndex of 0 so that
it can be focused, and then focuses it.
The call to focus is in render, since I couldn't figure out
how to place it in initialize, since not enough is loaded yet.
Now the dialog box closes when esc is pressed, like normal.

Fixes #196